test: improve Python sample fixtures
