+++ backend/tests/fixtures/sample_python_buggy.py
@@ -1,0 +1,17 @@
+import os
+
+password = "supersecret123"
+
+def calculate(a, b):
+    result = a / b
+    return result
+
+def risky():
+    try:
+        pass
+    except:
+        pass
+
+from os import *
+
+x = eval("1+2")
