+++ backend/tests/fixtures/sample_python_clean.py
@@ -1,0 +1,3 @@
+def add(a: int, b: int) -> int:
+    """Return the sum of a and b."""
+    return a + b
